% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{adaplasso}
\alias{adaplasso}
\title{Function for soft-thresholding}
\usage{
adaplasso(Xstd, Ystd, beta, lambda)
}
\arguments{
\item{Xstd}{n x p design matrix X scaled according to LARS algorithm and centered to mean 0}

\item{Ystd}{n x 1 centered output vector}

\item{beta}{p x 1 vector of parameters}

\item{lambda}{tuning parameter(scalar)}
}
\value{
Objective function for adaplasso
}
\description{
Function for soft-thresholding
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
#Standardizing X and Y
std <- standardize(X , Y , gamma)
#Deriving weighted and centered design matrix
Xstd <- std$Xstd
#Deriving centered Y
Ystd <- std$Ystd
#Defining beta
beta <- solve(crossprod(X)) \%*\% t(X) \%*\% Y
#Lambda value
lambda <- 2
#Objective function
obj <- adaplasso(Xstd, Ystd, beta, lambda)
}
