% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{adaplassostd_lambda}
\alias{adaplassostd_lambda}
\title{Fits adaptive adaplasso based on standardized data}
\usage{
adaplassostd_lambda(Xstd, Ystd, lambda, beta_init = NULL, eps = 0.001)
}
\arguments{
\item{Xstd}{n x p design matrix X scaled according to LARS algorithm and centered to mean 0}

\item{Ystd}{n x 1 centered output vector}

\item{lambda}{tuning parameter(scalar)}

\item{beta_init}{p x 1, optional starting point for coordinate descent algorithm}

\item{eps}{precision level for convergence assessment, default 0.001}
}
\value{
\item{beta}{vector of parameters}
\item{obj_min}{optimal value of the objective function}
}
\description{
Fits adaptive adaplasso based on standardized data
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
#Standardizing X and Y
std <- standardize(X , Y , gamma)
#Deriving weighted and centered design matrix
Xstd <- std$Xstd
#Deriving centered Y
Ystd <- std$Ystd
# tuning parameter
lambda  <- 0.1
fit <- adaplassostd_lambda(Xstd, Ystd, lambda)
}
