% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{adaplassostdseq_lambda}
\alias{adaplassostdseq_lambda}
\title{Fits Adaptive adaplasso on a sequence of lambda values based on standardized data}
\usage{
adaplassostdseq_lambda(
  Xstd,
  Ystd,
  tuning_seq = NULL,
  len_tuning = 60,
  eps = 0.001
)
}
\arguments{
\item{Xstd}{n x p design matrix X scaled according to LARS algorithm and centered to mean 0}

\item{Ystd}{n x 1 centered output vector}

\item{tuning_seq}{(optional)sequence of tuning parameters}

\item{len_tuning}{length of desired tuning parameter sequence}

\item{eps}{precision level for convergence assessment, default 0.001}
}
\value{
\item{tuning_seq}{the actual sequence of tuning parameters used}
\item{beta_lamb}{matrix of solutions at each lambda value for a given gamma, dimension is p x len_tuning
\item{obj_min_vec}{vector of optimal values of the objective function for each lambda at solution}
}
\description{
Fits Adaptive adaplasso on a sequence of lambda values based on standardized data
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
#Standardizing X and Y
std <- standardize(X , Y , gamma)
#Deriving weighted and centered design matrix
Xstd <- std$Xstd
#Deriving centered Y
Ystd <- std$Ystd
fit <- adaplassostdseq_lambda(Xstd, Ystd)
}
