% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{cv.gamma}
\alias{cv.gamma}
\title{Title Cross-Validation to choose the optimal gamma from a sequence of gamma values for a particular sequence of lambdas}
\usage{
cv.gamma(
  X,
  Y,
  lambda_seq = NULL,
  n_lambda = 60,
  gamma_seq = NULL,
  n_gamma = 60,
  k = 5,
  fold_ids = NULL,
  eps = 0.001
)
}
\arguments{
\item{X}{n x p matrix of covariates}

\item{Y}{n x 1 response vector}

\item{lambda_seq}{(optional)sequence of tuning parameters}

\item{n_lambda}{length of desired tuning parameter sequence, is only used when the tuning sequence is not supplied by the user}

\item{gamma_seq}{(optional)sequence of gamma values(used in determining weights)}

\item{n_gamma}{length of the desired sequence of gamma values}

\item{k}{number of folds for k-fold cross-validation, default is 5}

\item{fold_ids}{(optional) vector of length n specifying the folds assignment (from 1 to max(folds_ids)), if supplied the value of k is ignored}

\item{eps}{precision level for convergence assessment, default 0.001}
}
\value{
cvm = a n_gamma x n_lambda matrix giving CV(lambda, gamma) for each pair of (lambda, gamma), gamma_min = optimal gamma
}
\description{
Title Cross-Validation to choose the optimal gamma from a sequence of gamma values for a particular sequence of lambdas
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
fit_cv_gamma <- cv.gamma(X , Y)
}
