% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{cv.lambda}
\alias{cv.lambda}
\title{Perform cross-validation to select the best fit and finds the optimal lambda for a particular gamma value}
\usage{
cv.lambda(
  X,
  Y,
  tuning_seq = NULL,
  len_tuning = 60,
  gamma = 0.01,
  k = 5,
  id_fold = NULL,
  eps = 0.001
)
}
\arguments{
\item{X}{n x p  design matrix of inputs}

\item{Y}{n x 1 vector of outputs}

\item{tuning_seq}{(optional)sequence of tuning parameters}

\item{len_tuning}{length of desired tuning parameter sequence}

\item{gamma}{a scalar(>0) input used in the weight(user input)}

\item{k}{number of folds for k-fold cross-validation, default is 5}

\item{id_fold}{(optional) vector of length n specifying the folds assignment (from 1 to max(folds_ids)), if supplied the value of k is ignored}

\item{eps}{precision level for convergence assessment, default 0.001}
}
\value{
\item{tuning_seq}{the actual sequence of tuning parameters used}
\item{beta_lamb}{p x length(tuning_seq) matrix of corresponding solutions at each lambda value (original data without center or scale)}
\item{intercept_vec}{Unscaled vector of intercepts for a fixed gamma and for different lambda values}
\item{id_fold}{used splitting into folds from 1 to k (either as supplied or as generated in the beginning)}
\item{lambda_min}{selected lambda based on minimal rule}
\item{cv}{values of CV(lambda) for each lambda}
}
\description{
Perform cross-validation to select the best fit and finds the optimal lambda for a particular gamma value
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
fit_cv <- cv.lambda(X, Y)
}
