% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{cvadapLASSO}
\alias{cvadapLASSO}
\title{Perform cross-validation to select the best fit and finds the optimal lambda for a particular gamma value}
\usage{
cvadapLASSO(
  X,
  Y,
  lambda_seq = NULL,
  n_lambda = 60,
  gamma = 0.01,
  k = 5,
  fold_ids = NULL,
  eps = 0.001
)
}
\arguments{
\item{X}{n x p  design matrix of inputs}

\item{Y}{n x 1 vector of outputs}

\item{lambda_seq}{(optional)sequence of tuning parameters}

\item{n_lambda}{length of desired tuning parameter sequence, is only used when the tuning sequence is not supplied by the user}

\item{gamma}{a scalar(>0) input used in the weight(user input)}

\item{k}{number of folds for k-fold cross-validation, default is 5}

\item{fold_ids}{(optional) vector of length n specifying the folds assignment (from 1 to max(folds_ids)), if supplied the value of k is ignored}

\item{eps}{precision level for convergence assessment, default 0.001}
}
\value{
\item{lambda_seq}{the actual sequence of tuning parameters used}
\item{beta_mat}{p x length(lambda_seq) matrix of corresponding solutions at each lambda value (original data without center or scale)}
\item{beta0_vec}{length(lambda_seq) vector of intercepts (original data without center or scale)}
\item{fold_ids}{used splitting into folds from 1 to k (either as supplied or as generated in the beginning)}
\item{lambda_min}{selected lambda based on minimal rule}
\item{lambda_1se}{selected lambda based on 1SE rule}
\item{cvm}{values of CV(lambda) for each lambda}
\item{cvse}{values of SE_CV(lambda) for each lambda}
}
\description{
Perform cross-validation to select the best fit and finds the optimal lambda for a particular gamma value
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
fit_cv <- cvadapLASSO(X, Y)
}
