% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{fitadapLASSOstandardized}
\alias{fitadapLASSOstandardized}
\title{Title}
\usage{
fitadapLASSOstandardized(
  Xtilde,
  Ytilde,
  lambda,
  beta_start = NULL,
  eps = 0.001
)
}
\arguments{
\item{Xtilde}{n x p matrix, centered and scaled X where X has been scaled as mentioned in LARS algorithm}

\item{Ytilde}{n x 1 vector of centered outputs}

\item{lambda}{tuning parameter}

\item{beta_start}{p x 1, optional starting point for coordinate descent algorithm}

\item{eps}{precision level for convergence assessment, default 0.001}
}
\value{
beta = vector of parameters , fmin = optimal value of the objective function
}
\description{
Title
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
#Standardizing X and Y
std <- standardizeXY(X , Y , gamma)
#Deriving weighted and centered design matrix
Xtilde <- std$Xtilde
#Deriving centered Y
Ytilde <- std$Ytilde
# tuning parameter
lambda  <- 0.1
fit <- fitadapLASSOstandardized(Xtilde, Ytilde, lambda)
}
