% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{fitadapLASSOstandardized_seq}
\alias{fitadapLASSOstandardized_seq}
\title{Fits Adaptive Lasso on a sequence of lambda values based on standardized data}
\usage{
fitadapLASSOstandardized_seq(
  Xtilde,
  Ytilde,
  lambda_seq = NULL,
  n_lambda = 60,
  eps = 0.001
)
}
\arguments{
\item{Xtilde}{n x p design matrix X scaled according to LARS algorithm and centered to mean 0}

\item{Ytilde}{n x 1 centered output vector}

\item{lambda_seq}{(optional)sequence of tuning parameters}

\item{n_lambda}{length of desired tuning parameter sequence, is only used when the tuning sequence is not supplied by the user}

\item{eps}{precision level for convergence assessment, default 0.001}
}
\value{
\item{lambda_seq}{the actual sequence of tuning parameters used}
\item{beta_mat}{p x length(lambda_seq) matrix of corresponding solutions at each lambda value}
\item{fmin_vec}{length(lambda_seq) vector of corresponding objective function values at solution}
}
\description{
Fits Adaptive Lasso on a sequence of lambda values based on standardized data
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
#Standardizing X and Y
std <- standardizeXY(X , Y , gamma)
#Deriving weighted and centered design matrix
Xtilde <- std$Xtilde
#Deriving centered Y
Ytilde <- std$Ytilde
fit <- fitadapLASSOstandardized_seq(Xtilde, Ytilde)
}
