% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{lasso}
\alias{lasso}
\title{Function for soft-thresholding}
\usage{
lasso(Xtilde, Ytilde, beta, lambda)
}
\arguments{
\item{Xtilde}{n x p design matrix X scaled according to LARS algorithm and centered to mean 0}

\item{Ytilde}{n x 1 centered output vector}

\item{beta}{p x 1 vector of parameters}

\item{lambda}{tuning parameter(scalar)}
}
\value{
Objective function for Lasso
}
\description{
Function for soft-thresholding
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
#Standardizing X and Y
std <- standardizeXY(X , Y , gamma)
#Deriving weighted and centered design matrix
Xtilde <- std$Xtilde
#Deriving centered Y
Ytilde <- std$Ytilde
#Defining beta
beta <- solve(crossprod(X)) \%*\% t(X) \%*\% Y
#Lambda value
lambda <- 2
#Objective function
obj <- lasso(Xtilde, Ytilde, beta, lambda)
}
