% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{scale_X}
\alias{scale_X}
\title{scale_X Scales a matrix of inputs according to  LARS algorithm}
\usage{
scale_X(X, Y, gamma)
}
\arguments{
\item{X}{n x p  design matrix of inputs}

\item{Y}{n x 1 vector of outputs}

\item{gamma}{a scalar(>0) input used in the weight(user input)}
}
\value{
A list with the elements
\item{X_w}{A n x p matrix scaled according to LARS algorithm}
\item{weights}{adaptive weights}
}
\description{
scale_X Scales a matrix of inputs according to  LARS algorithm
}
\examples{
EXAMPLE
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
#Scaling using scale_X
sc <- scale_X(X , Y , gamma)
#Deriving weighted design matrix
X_w <- sc$X_w
#Deriving weights
weights <- sc$weights
}
