% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{standardize}
\alias{standardize}
\title{Standardizes the input design matrix X and output vector Y to mean 0 and scales X}
\usage{
standardize(X, Y, gamma)
}
\arguments{
\item{X}{n x p  design matrix of inputs}

\item{Y}{n x 1 vector of outputs}

\item{gamma}{a scalar(>0) input used in the weight(user input)}
}
\value{
\item{Xstd}{scaled X}
\item{Ystd}{scaled Y}
\item{meanY}{mean of Y}
\item{meanX}{Column means after centering the weighted X matrix from scale_X}
\item{weights}{weights obtained by centering X_w which is obtained from scale X}
}
\description{
Standardizes the input design matrix X and output vector Y to mean 0 and scales X
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
#Standardizing X and Y
std <- standardize(X , Y , gamma)
#Deriving weighted and centered design matrix
Xstd <- std$Xstd
#Column means of centered X_w
meanX <- std$meanX
#Deriving centered Y
Ystd <- std$Ystd
#Mean of Ystd
meanY <- std$meanY
# Weights
weights <- std$weights
}
