% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{standardizeXY}
\alias{standardizeXY}
\title{Standardizes the input design matrix X and output vector Y to mean 0 and scales X}
\usage{
standardizeXY(X, Y, gamma)
}
\arguments{
\item{X}{n x p  design matrix of inputs}

\item{Y}{n x 1 vector of outputs}

\item{gamma}{a scalar(>0) input used in the weight(user input)}
}
\value{
\item{Xtilde}{scaled X}
\item{Ytilde}{scaled Y}
\item{Ymeans}{mean of Y}
\item{Xmeans}{Column means after centering the weighted X matrix from scale_X}
\item{weights}{weights obtained by centering X_w which is obtained from scale X}
}
\description{
Standardizes the input design matrix X and output vector Y to mean 0 and scales X
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
#Standardizing X and Y
std <- standardizeXY(X , Y , gamma)
#Deriving weighted and centered design matrix
Xtilde <- std$Xtilde
#Column means of centered X_w
Xmeans <- std$Xmeans
#Deriving centered Y
Ytilde <- std$Ytilde
#Mean of Ytilde
Ymean <- std$Ymean
# Weights
weights <- std$weights
}
