% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{cv.gamma}
\alias{cv.gamma}
\title{Cross-Validation to choose the optimal gamma from a sequence of gamma values for a particular sequence of lambdas}
\usage{
cv.gamma(
  X,
  Y,
  tuning_seq = NULL,
  len_tuning = 60,
  gamma_seq = NULL,
  n_gamma = 60,
  k = 5,
  id_fold = NULL,
  eps = 0.001
)
}
\arguments{
\item{X}{n x p  design matrix of inputs}

\item{Y}{n x 1 vector of outputs}

\item{tuning_seq}{(optional)sequence of tuning parameters}

\item{len_tuning}{length of desired tuning parameter sequence}

\item{gamma_seq}{(optional)sequence of gamma values(used in determining weights)}

\item{n_gamma}{length of the desired sequence of gamma values}

\item{k}{number of folds for k-fold cross-validation, default is 5}

\item{id_fold}{(optional) vector of length n specifying the folds assignment (from 1 to max(folds_ids)), if supplied the value of k is ignored}

\item{eps}{precision level for convergence assessment, default 0.001}
}
\value{
\item{cv}{a n_gamma x len_tuning matrix giving CV(lambda, gamma) for each pair of (lambda, gamma)}
\item{gamma_min}{optimal gamma}
\item{lambda_min}{selected lambda based on minimal rule}
}
\description{
Cross-Validation to choose the optimal gamma from a sequence of gamma values for a particular sequence of lambdas
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
fit_cv_gamma <- cv.gamma(X , Y)
}
