% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{fitadapLASSO}
\alias{fitadapLASSO}
\title{Fits adaptive LASSO}
\usage{
fitadapLASSO(X, Y, lambda_seq = NULL, n_lambda = 60, gamma = 0.01, eps = 0.001)
}
\arguments{
\item{X}{n x p  design matrix of inputs}

\item{Y}{n x 1 vector of outputs}

\item{lambda_seq}{(optional)sequence of tuning parameters}

\item{n_lambda}{length of desired tuning parameter sequence, is only used when the tuning sequence is not supplied by the user}

\item{gamma}{a scalar(>0) input used in the weight(user input)}

\item{eps}{precision level for convergence assessment, default 0.001}
}
\value{
lambda_seq = the actual sequence of tuning parameters used,beta_mat = p x length(lambda_seq) matrix of corresponding solutions at each lambda value (original data without center or scale), beta0_vec = length(lambda_seq) vector of intercepts (original data without center or scale)
}
\description{
Fits adaptive LASSO
}
\examples{
#EXAMPLE 1
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
# Fits adaptive LASSO
fit <- fitadapLASSO(X , Y , gamma = gamma)
# EXAMPLE 2
X <- matrix(rchisq(500, 3), 50, 10)
Y <- rbinom(50)
lambda_seq <- runif(100, 1, 2)
#Fits adaptive lasso
fit2 <- fitadapLASSO(X, Y, lambda_seq = lambda_seq, gamma = 0.1, eps = 0.002 )
}
