% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{fitadaplasso}
\alias{fitadaplasso}
\title{Fits adaptive lasso}
\usage{
fitadaplasso(
  X,
  Y,
  tuning_seq = NULL,
  len_tuning = 60,
  gamma = 0.01,
  eps = 0.001
)
}
\arguments{
\item{X}{n x p  design matrix of inputs}

\item{Y}{n x 1 vector of outputs}

\item{tuning_seq}{(optional)sequence of tuning parameters}

\item{len_tuning}{length of desired tuning parameter sequence}

\item{gamma}{a scalar(>0) input used in the weight(user input)}

\item{eps}{precision level for convergence assessment, default 0.001}
}
\value{
\item{tuning_seq}{the actual sequence of tuning parameters used}
\item{beta_lamb}{p x length(tuning_seq) matrix of corresponding solutions at each lambda value (original data without center or scale)}
\item{intercept_vec} {Unscaled vector of intercepts for a fixed gamma and for different lambda values}
}
\description{
Fits adaptive lasso
}
\examples{
#EXAMPLE 1
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
# Fits adaptive adaplasso
fit <- fitadaplasso(X , Y , gamma = gamma)
# EXAMPLE 2
X <- matrix(rchisq(500, 3), 50, 10)
Y <- rbinom(50)
tuning_seq <- runif(100, 1, 2)
#Fits adaptive adaplasso
fit2 <- fitadaplasso(X, Y, tuning_seq = tuning_seq, gamma = 0.1, eps = 0.002 )
}
