% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{fitadapLASSOstandardized}
\alias{fitadapLASSOstandardized}
\title{Fits adaptive lasso based on standardized data}
\usage{
fitadapLASSOstandardized(
  Xtilde,
  Ytilde,
  lambda,
  beta_start = NULL,
  eps = 0.001
)
}
\arguments{
\item{Xtilde}{n x p design matrix X scaled according to LARS algorithm and centered to mean 0}

\item{Ytilde}{n x 1 centered output vector}

\item{lambda}{tuning parameter(scalar)}

\item{beta_start}{p x 1, optional starting point for coordinate descent algorithm}

\item{eps}{precision level for convergence assessment, default 0.001}
}
\value{
\item{beta}{vector of parameters}
\item{fmin}{optimal value of the objective function}
}
\description{
Fits adaptive lasso based on standardized data
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
#Standardizing X and Y
std <- standardizeXY(X , Y , gamma)
#Deriving weighted and centered design matrix
Xtilde <- std$Xtilde
#Deriving centered Y
Ytilde <- std$Ytilde
# tuning parameter
lambda  <- 0.1
fit <- fitadapLASSOstandardized(Xtilde, Ytilde, lambda)
}
