% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{fitadapLASSOstandardized_seq}
\alias{fitadapLASSOstandardized_seq}
\title{Title fits Adaptive Lasso on a sequence of lambda values based on standardized data}
\usage{
fitadapLASSOstandardized_seq(
  Xtilde,
  Ytilde,
  lambda_seq = NULL,
  n_lambda = 60,
  eps = 0.001
)
}
\arguments{
\item{Xtilde}{n x p matrix, centered and scaled X where X has been scaled as mentioned in LARS algorithm}

\item{Ytilde}{n x 1 vector of centered outputs}

\item{lambda_seq}{(optional)sequence of tuning parameters}

\item{n_lambda}{length of desired tuning parameter sequence, is only used when the tuning sequence is not supplied by the user}

\item{eps}{precision level for convergence assessment, default 0.001}
}
\value{
lambda_seq=the actual sequence of tuning parameters used,beta_mat=p x length(lambda_seq) matrix of corresponding solutions at each lambda value,fmin_vec = fmin_vec - length(lambda_seq) vector of corresponding objective function values at solution
}
\description{
Title fits Adaptive Lasso on a sequence of lambda values based on standardized data
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
#Standardizing X and Y
std <- standardizeXY(X , Y , gamma)
#Deriving weighted and centered design matrix
Xtilde <- std$Xtilde
#Deriving centered Y
Ytilde <- std$Ytilde
fit <- fitadapLASSOstandardized_seq(Xtilde, Ytilde)
}
