% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{soft}
\alias{soft}
\title{Title}
\usage{
soft(a, lambda)
}
\arguments{
\item{a}{scalar to be soft-thresholded}

\item{lambda}{level of soft thresholding}
}
\value{
soft-thresholded value
}
\description{
Title
}
\examples{
a = 2
lambda = 1
soft(a , lambda)
}
