% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FunctionsALasso.R
\name{standardizeXY}
\alias{standardizeXY}
\title{Title Standardizes the input design matrix X and output vector Y to mean 0 and scales X}
\usage{
standardizeXY(X, Y, gamma)
}
\arguments{
\item{X}{n x p design matrix of inputs}

\item{Y}{n x 1 output vector}

\item{gamma}{a scalar (>0) input(user-defined) used in the weights}
}
\value{
Xtilde=scaled X, Ytilde= scaled Y, Ymeans = mean of Y , Xmeans =  Column means after centering the weighted X matrix from scale_X , weights = weights obtained by centering X_w which is obtained from scale X
}
\description{
Title Standardizes the input design matrix X and output vector Y to mean 0 and scales X
}
\examples{
X <- matrix(rnorm(500), 50, 10)
Y <- rnorm(50)
gamma <- 2
#Standardizing X and Y
std <- standardizeXY(X , Y , gamma)
#Deriving weighted and centered design matrix
Xtilde <- std$Xtilde
#Column means of centered X_w
Xmeans <- std$Xmeans
#Deriving centered Y
Ytilde <- std$Ytilde
#Mean of Ytilde
Ymean <- std$Ymean
# Weights
weights <- std$weights
}
